Fix/endpoints

+++ src/main/resources/schema.sql
@@ -57,7 +57,7 @@ CREATE TABLE IF NOT EXISTS pasillos (
 CREATE TABLE IF NOT EXISTS estanterias (
   id INT NOT NULL AUTO_INCREMENT,
   codigo CHAR(1) NOT NULL COMMENT 'Ejemplo: A',
-  niveles_maximos TINYINT NOT NULL,
+  niveles_maximos INT NOT NULL,
   capacidad_nivel INT NOT NULL DEFAULT 8 COMMENT 'En cada nivel cabe solo un palé con un maximo de 8 cajas',
   pasillo_id INT NOT NULL,
   PRIMARY KEY (id),
@@ -70,7 +70,7 @@ CREATE TABLE IF NOT EXISTS ubicaciones_almacen (
   id INT NOT NULL AUTO_INCREMENT,
   referencia VARCHAR(255) NOT NULL COMMENT 'Ejemplo: 1A3 (pasillo-estantería-nivel)',
   estanteria_id INT NOT NULL,
-  nivel TINYINT NOT NULL,
+  nivel INT NOT NULL,
   PRIMARY KEY (id),
   CONSTRAINT fk_ubicaciones_estanterias
     FOREIGN KEY (estanteria_id) REFERENCES estanterias(id),
